Room (1,0) end to end on the explicit graph: 6:33 vs 7:13, every ladder fingerprint identical
Three scale-only bugs fixed: empty pieces are dropped at finish (their
index-renumbered seq collided with a real one), units of nothing but won
lanes are not run (their shape has no kernel), and the row ref names the
queue in 24 bits (spares are per outcome, so the slot vector grows past
POOL_QUEUES and an index >= 256 aliased another queue's row: 29k spurious
marks). bench-backward --diff expands disputed states, lists the recorded
edges from them against their real successors, and (CELESTE_DIFF_RERUN)
re-runs the frame with the tree's door.

+++ src/compiled/asm_kernel.rs
@@ -342,7 +342,7 @@ impl AsmKernel {
                             // The row was flushed (a stale ref, only if
                             // the cache lost the flush's write-back): push
                             // it again; the flush merges duplicates by key.
-                            Some(b) if !sink.mark_pred(r as u32, b, i) => {}
+                            Some(b) if !sink.mark_pred(r, b, i) => {}
                             _ => continue,
                         }
                     }
@@ -362,7 +362,7 @@ impl AsmKernel {
                         sink.slots[q].pred_base.push(b);
                         sink.slots[q].pred_mask.push(1 << i);
                         sink.slots[q].last_extra.push(u32::MAX);
-                        sink.seen.set_ref(key, sink.row_ref(q) as u64);
+                        sink.seen.set_ref(key, sink.row_ref(q));
                     }
                     sink.pushed(q).expect("flushing a full queue");
                 }
